requirements.txt
pandas==2.0.3
mysql-connector-python==8.0.33
apache-airflow==2.6.3
python-dotenv==1.0.0
